// src/generated/css/CssDeclaration.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';

/**
 * CssDeclaration
 * @generated
 */
export interface CssDeclaration extends EObject {
  property: string;
  value: string;
  important?: boolean;

}